backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymongo==4.9.2
motor==3.6.0
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.16
email-validator==2.2.0
python-dateutil==2.9.0
geopy==2.4.1
stripe==11.5.0
sendgrid==6.11.0
twilio==9.3.2
redis==5.1.1
celery==5.4.0
httpx==0.28.1
oauthlib==3.2.2
pytest==8.3.4
pytest-asyncio==0.21.1
pytest-mock==3.11.1
pytest-cov==4.1.0
python-json-logger==2.0.7
prometheus-client==0.23.0
sentry-sdk[fastapi]==2.21.0

# AI/ML Dependencies
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-anthropic==0.3.0
huggingface-hub==0.27.0
chromadb==0.5.23
sentence-transformers==3.3.1
transformers==4.48.0
torch==2.5.1
numpy==2.2.1
pandas==2.2.3
scikit-learn==1.6.0
spacy==3.8.3
nltk==3.9.1
faiss-cpu==1.9.0.post1
tiktoken==0.8.0
openai==1.58.1
anthropic==0.42.0
google-generativeai==0.8.3
pinecone-client==5.0.1
qdrant-client==1.12.1
weaviate-client==4.10.1

# Spacy language model (install after spacy)
# Run: python -m spacy download en_core_web_sm

# Kafka dependencies
kafka-python==2.0.2
confluent-kafka==2.7.0
avro-python3==1.10.2
